udh digabung css sebagian sama imathlatest

+++ imathivi/application/controllers/admin/pesan.php
@@ -2,18 +2,25 @@
     class pesan extends CI_Controller{
         public function index()
         {
-            $this->load->model('pesan_model');
- 
-            $data['result'] = $this->pesan_model->getAllPesan();	    
-	    
-	    $this->load->view('admin/pesan_view',$data);
+		if($this->session->userdata('role')=="admin") {
+			$this->load->model('pesan_model');	 
+			$data['result'] = $this->pesan_model->getAllPesan();	    
+			$this->load->view('admin/pesan_view',$data);
+		} 
+		else {
+			redirect('home');
+		}	
 	}
 	
 	public function delete($id){
-		$this->load->model('pesan_model');
-		
-		$this->pesan_model->delete($id);
-		redirect('admin/pesan', 'refresh');
+		if($this->session->userdata('role')=="admin") {
+			$this->load->model('pesan_model');
+			$this->pesan_model->delete($id);
+			redirect('admin/pesan', 'refresh');
+		} 
+		else {
+			redirect('home');
+		}	
 	}
 	
     }
